noticed that Jennifer and I's implementation are a bit conflicting. Needs to be resolved. Also, resolved Shenyang's review

+++ src/main/java/use_case/loyalty_score/LoyaltyScoringService.java
@@ -2,7 +2,9 @@ package use_case.loyalty_score;
 
 /**
  * A wrapper class for calculating loyalty score.
+ * Currently conflicting with Jennifer's loyalty score algorithm
  */
+
 public class LoyaltyScoringService {
 
     private static final double K = 0.07;
@@ -10,7 +12,7 @@ public class LoyaltyScoringService {
      * Calculates score, 1 to 100 from rank in top 50 items. 1st ~ 100 loyalty, 50th ~ 3.
      */
 
-    public static double scoreFromRank(int rank) {
+    public static int scoreFromRank(int rank) {
         double unrounded = 100 * Math.exp(-K * (rank - 1));
         return (int) Math.round(unrounded);
     }
